clean up riff code
