+++ OpenLocoTool/DatFileParsing/SawyerStreamReader.cs
@@ -7,57 +7,6 @@ using OpenLocoToolCommon;
 
 namespace OpenLocoTool.DatFileParsing
 {
-	[LocoStructSize(0x2C)]
-	public record MusicHeader(
-		[property: LocoStructOffset(0x00)] uint32_t Signature,
-		[property: LocoStructOffset(0x04)] uint32_t Size,
-		[property: LocoStructOffset(0x08)] uint32_t RiffType,
-		[property: LocoStructOffset(0x0C)] uint32_t FormatMarker,
-		[property: LocoStructOffset(0x10)] uint32_t HeaderSize,
-		[property: LocoStructOffset(0x14)] uint16_t FormatType,
-		[property: LocoStructOffset(0x16)] uint16_t NumberOfChannels,
-		[property: LocoStructOffset(0x18)] uint32_t SampleRate,
-		[property: LocoStructOffset(0x1C)] uint32_t unk1,
-		[property: LocoStructOffset(0x20)] uint16_t unk2,
-		[property: LocoStructOffset(0x22)] uint16_t BitsPerSample,
-		[property: LocoStructOffset(0x24)] uint32_t DataMarker,
-		[property: LocoStructOffset(0x28)] uint32_t PcmDataLength) : ILocoStruct
-	{
-		public bool Validate()
-		{
-			if (Signature != 0x46464952) // "RIFF"
-			{
-				return false;
-			}
-
-			if (RiffType != 0x45564157) // "WAVE"
-			{
-				return false;
-			}
-
-			if (FormatMarker != 0x20746d66 && FormatMarker != 0x00746d66) // "fmt\0" or "fmt"
-			{
-				return false;
-			}
-
-			if (FormatType != 1) // expected PCM
-			{
-				return false;
-			}
-
-			if (BitsPerSample != 16)
-			{
-				return false;
-			}
-
-			if (DataMarker != 0x61746164)
-			{
-				return false;
-			}
-
-			return true;
-		}
-	}
 
 	public static class SawyerStreamReader
 	{
@@ -419,15 +368,15 @@ namespace OpenLocoTool.DatFileParsing
 			_ => throw new InvalidDataException("Unknown chunk encoding scheme"),
 		};
 
-		public static (MusicHeader header, byte[] data) LoadMusicTrack(byte[] data)
+		public static (RiffWavHeader header, byte[] data) LoadMusicTrack(byte[] data)
 		{
 			using (var ms = new MemoryStream(data))
 			using (var br = new BinaryReader(ms))
 			{
-				var headerBytes = br.ReadBytes(ObjectAttributes.StructSize<MusicHeader>());
-				var header = ByteReader.ReadLocoStruct<MusicHeader>(headerBytes);
+				var headerBytes = br.ReadBytes(ObjectAttributes.StructSize<RiffWavHeader>());
+				var header = ByteReader.ReadLocoStruct<RiffWavHeader>(headerBytes);
 
-				var pcmData = new byte[header.PcmDataLength];
+				var pcmData = new byte[header.DataLength];
 				br.Read(pcmData);
 				return (header, pcmData);
 			}
